Addressing various feedback items and fixing lint warnings.
