+++ lib/src/memory/memories.dart
@@ -1,7 +1,9 @@
 // Copyright (C) 2023-2024 Intel Corporation
 // SPDX-License-Identifier: BSD-3-Clause
 
-export 'csr.dart';
-export 'csr_config.dart';
+export 'csr/csr.dart';
+export 'csr/csr_block.dart';
+export 'csr/csr_config.dart';
+export 'csr/csr_top.dart';
 export 'memory.dart';
 export 'register_file.dart';
